H2 angepasst

+++ src/main/kotlin/oemer/aytteamforge/controller/dto/in/TeamsRequestDto.kt
@@ -6,4 +6,4 @@ import com.fasterxml.jackson.annotation.JsonProperty
 data class TeamsRequestDto @JsonCreator constructor(
     @JsonProperty("players") val players: List<PlayerDto>)
 
-data class PlayerDto(val name: String, val skill: Double?)
+data class PlayerDto(val name: String, val skill: Int?)
